Fixed include and name mangling bug

+++ include/Config.hpp
@@ -7,7 +7,7 @@ namespace ic::config
 {
     constexpr uint16_t DEFAULT_PORT = 15317;
     constexpr uint8_t ISENCOIN_VERSION = 1;
-    constexpr uint8_t ISENCOIN_REWARD = 10;
+    constexpr float ISENCOIN_REWARD = 10;
     constexpr std::array<unsigned char, 32> ISENCOIN_NULL_ADDRESS = {};
     namespace tx_field_chars
     {
